Further divide into smaller separate file. The main.rs is too large.
